Refactor chain, removing static_validate_header and validate_transaction and instead dispatching via traits
