--- category/execution/ethereum/chain/chain.cpp
@@ -1,33 +1,0 @@
-// Copyright (C) 2025 Category Labs, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-#include <category/execution/ethereum/chain/chain.hpp>
-
-#include <category/core/config.hpp>
-#include <category/core/result.hpp>
-
-#include <boost/outcome/config.hpp>
-#include <boost/outcome/success_failure.hpp>
-
-MONAD_NAMESPACE_BEGIN
-
-using BOOST_OUTCOME_V2_NAMESPACE::success;
-
-Result<void> Chain::static_validate_header(BlockHeader const &) const
-{
-    return success();
-}
-
-MONAD_NAMESPACE_END
