+++ category/execution/ethereum/chain/ethereum_mainnet.hpp
@@ -43,15 +43,7 @@ struct EthereumMainnet : Chain
     virtual evmc_revision
     get_revision(uint64_t block_number, uint64_t timestamp) const override;
 
-    virtual Result<void>
-    static_validate_header(BlockHeader const &) const override;
-
     virtual GenesisState get_genesis_state() const override;
-
-    virtual Result<void> validate_transaction(
-        uint64_t block_number, uint64_t timestamp, Transaction const &,
-        Address const &sender, State &, uint256_t const &base_fee_per_gas,
-        std::span<std::optional<Address> const> authorities) const override;
 };
 
 MONAD_NAMESPACE_END
