+++ category/execution/ethereum/transaction_gas.hpp
@@ -40,10 +40,6 @@ template <Traits traits>
 uint256_t
 gas_price(Transaction const &, uint256_t const &base_fee_per_gas) noexcept;
 
-uint256_t gas_price(
-    evmc_revision, Transaction const &,
-    uint256_t const &base_fee_per_gas) noexcept;
-
 template <Traits traits>
 uint64_t g_star(Transaction const &, uint64_t gas_remaining, uint64_t refund);
