+++ cmd/monad/runloop_monad.cpp
@@ -187,7 +187,6 @@ Result<BlockExecOutput> propose_block(
 
     // Block input validation
     BOOST_OUTCOME_TRY(static_validate_consensus_header(consensus_header));
-    BOOST_OUTCOME_TRY(chain.static_validate_header(block.header));
     BOOST_OUTCOME_TRY(static_validate_block<traits>(chain, block));
 
     // Sender and EIP-7702 authorities recovery
