+++ cmd/monad/runloop_monad_ethblocks.cpp
@@ -125,7 +125,6 @@ Result<void> process_monad_block(
     auto const block_begin = std::chrono::steady_clock::now();
 
     // Block input validation
-    BOOST_OUTCOME_TRY(chain.static_validate_header(block.header));
     BOOST_OUTCOME_TRY(static_validate_block<traits>(chain, block));
 
     // Sender and authority recovery
